arm64: dts: lahaina/yupik: Remove display ramdump memory region
This region is the same as the early splash framebuffer initialized by
XBL. msm_drm uses it for ramdump when available, but that's a debugging
feature we don't need. It also wastes memory after early boot because
msm_drm will keep it reserved rather than freeing the framebuffer memory
once it has taken control of the display hardware.

Remove the display ramdump region to save 36 MiB of memory.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Change-Id: Ib20eb14022e81c1c91ab72263ece67794f23cf26
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Signed-off-by: TogoFire <togofire@mailfence.com>