lpm-levels: Treat negative sleep lengths as indefinite
A negative sleep length means there's no hrtimer armed, and therefore the
sleep length is indefinite if the tick is stopped. Fix this logic in TEO to
allow a deeper idle state to correctly be used when there's no hrtimer
armed and thus the CPU won't be prematurely woken up from its sleep.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I8e9781e77d6c0759360ae499fc7410845891eac1
[dereference23: Port to lpm-levels]
Signed-off-by: Alexander Winkowski <dereference23@outlook.com>
Signed-off-by: TogoFire <togofire@mailfence.com>