Revert "drivers: oplus_sensor_devinfo: Increase SOURCE_NUM to 3"
This reverts commit 261da90465e90fd0d8603b71b1b248a232a92173.

Because it is not on the most current firmware at the moment and is using the older a14. This will not be compatible. So do without it for now. Therefore, the touch does not work, the battery does not charge, video and audio do not play.

And btw:
[Vishalcj17]:
Ask users to use OOS 14.0.0.202/14.0.0.212 firmware instead.
14.0.0.500+ firmware causes bootloader breakage.

Signed-off-by: TogoFire <togofire@mailfence.com>
Change-Id: Iba1973237b4584b7cbeef6c1bd655b632b53d03e