sched: drop dvfs headroom
Revert "sched: Implement tapered dvfs_headroom"
This reverts commit 8edbf28945651eca001b659a16e877e240f4a5b2.

Revert "sched: cpufreq: Apply DVFS headroom to CFS only"
This reverts commit 5a264df48eac178169763c6bbe011e3b62f92d48.

Reason:
This was a backport, which is incomplete and unstable. So we have to remove it.
This is both bad and good at the same time. It gives you great battery life, especially using 4g or 5g, but it causes a drop in fps. Although in theory it is supposed to have good performance.

What is headroom?
When cpu util is low, it gets artificially inflated (gets headroom). Non-constant - the closer util to peak cpu capacity - the less inflation it gets. This thing helps with certain kinds of tasks, especially ui, without ruining battery life much.
Util boost is applied to summed util per cpu (with or without RT and deadline tasks, irq, whatever you set in code), not per-task.
DVFS stands for Dynamic Voltage and Frequency Scaling. Basically, it is a technique that allows the processor to adjust its speed and power consumption according to the workload. "Headroom" is like an extra performance space that the system reserves to handle peaks in demand.

How does headroom affect FPS?
* High headroom: It can cause the processor to maintain a higher frequency than necessary, consuming more power and potentially generating more heat. This can cause FPS drops if the device is subjected to thermal throttling (reducing performance to avoid overheating).
* Low headroom: It can cause the processor to take longer to increase the frequency when necessary, causing slowdowns and FPS drops in moments of high demand.
* Variable headroom: The implementation discussed in the conversation (tapered headroom) tries to find a balance between the two scenarios above, adjusting the headroom according to the current CPU utilization. This can be beneficial in some cases.

You could increase the value to reach some middle ground to see if it becomes a little more stable, but this is not very viable, better not to change it. For example:
Start with a 20% increase: If the default value is 1280, try increasing it to 1536.
Don't go above 50%: Increasing the value beyond 50% can lead to instability and damage the device.

We could also try these commits, but they are bad and unstable, not worth it.
-sched: Introduce Per-Cluster DVFS Headroom
https://github.com/EmanuelCN/kernel_xiaomi_sm8250/commit/ca0ebeab178b2b568c913b4ed0ea5bbe6585c4da
* Map util perf is a headroom, which he did only for the cluster. But that's not good. Headroom hax aren't a good solution. The scheduler has no notion of that. Although the scheduler is designed to cope with the CPU running at a different speed than it expects, it throws a wrench into everything. There will be lag.
-sched/cpufreq: Rework schedutil governor performance estimation https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit?id=9c0b4bb7f6303c9c4e2e34984c46f5a86478f84d https://github.com/EmanuelCN/kernel_xiaomi_sm8250/commit/5ec131cfd0c0e4b78a045a948812d490170b07ef

Signed-off-by: TogoFire <togofire@mailfence.com>
Change-Id: I8bc493850e3bc7d45414fb9127c83308500dd300