Revert "techpack: disp: msm: Correct the panel's LP11_INIT"
This reverts commit 468b36f6a1cbbdb2a281b6ad1475b2b43ebd06e6.

The Iris will be reinstated. Therefore, this change must be undone as it's causing the pixelworks (Iris) to malfunction. This resolve the issue where the display doesn't turn on after locking and the BSOD errors.

Useless anyway. Given the negative impact on jitter, it's imperative that we revert this commit.

Change-Id: 0c76ccb6c9766f99de9c2530b51e41496a98a453
Signed-off-by: TogoFire <togofire@mailfence.com>